Don't reveal partial passwords
When typing a wifi password, if a substring of the password matched
the $PASSWORD_ENTER prompt, it would highlight the matching letters,
revealing a part of the password. Setting the prompt to be only spaces
should mitigate this.

The width of the prompt determines the width of the displayed box; thus
we cannot have an empty prompt, and must hope that no one in their right
mind makes a password that is simply several spaces in a row.